dashboard/requirements.txt
boto3
requests
pandas
psycopg2-binary
python-dotenv
streamlit
boto3
botocore
